#55 feat : 재도전 시 다음날짜로 자동 설정 기능 구현
